fix logging
